Don't set dispositionError to encoding.Error(nil) since it won't properly compare to (error)nil. (#72)
This fixes it so we don't propagate Error(nil) and just let dispositionError default to a error(nil).

Basically, if we had an error when rejecting the code works well. It returns a non-nil Error, and we expect the customer to look at it. However, if the `Error` was nil, then it won't properly compare to error(nil). So even when we succeeded it looked a bit like we failed. 

Part of a fix for a bug found in Service Bus: https://github.com/Azure/azure-service-bus-go/issues/215