Message properties that are optional need to be nil-able (#93)
Matching the spec on the message properties - all of them should have been marked optional, which means all of them should be 'ptr to type'. This specific problem made it so you couldn't pass in an empty session ID (which is valid) since it would assume you were trying to omit the field.

Also, added in some convenience type annotations to make it simpler to scan the types vs the spec and make sure they match.

NOTE: this is a breaking change for any clients if they are using any of the message properties fields.