Merge flowcontrol branch into main (#260)
* Replace messages channel with a queue (#244)

* Replace messages channel with a queue

The Receiver.messages channel placed an arbitrary restriction on the
amount of link credit that could be issued to a Receiver.  Once the
Receiver has been created, it can never be issued credit that exceeds
the size of the channel.
In addition, if the Receiver were to receive messages that exceeds the
size of the channel, due to flow control bugs or other, the writes to
the channel could block leading to hangs.
The channel has been replaced with a segmented FIFO queue.  While this
in theory could allow for unbounded growth, the reality is that the
total size can never be greater than that of the Session's incoming
window.

* use Ring from standard library

* switch to require.Same for pointer comparisons

fixed issue in last seg check and added another test

* refine and add another test

* size message queue to the incoming window

* add Holder[T] to abstract syncronized access

* Remove concept of "max credit" from Receivers (#251)

* Remove concept of "max credit" from Receivers

This exposed an implementation detail of how Receivers worked.  With the
removal of the Message channel, this is no longer required and never
really made sense anyways.
The MaxCredit and ManualCredits options have been consolidated into a
single Credit option.  This is used to set the total credit for a
Receiver in auto-flow mode, or to disable auto-flow.
Disposition batching needed some slight refactoring.  Now you can
directly specify the batch size instead of it being tied to link credit.
Received messages that are sender-settled now correctly count towards a
Receiver's credit and must be settled to reclaim it.

* fix linter

* back out double-buffering change for now

* refine threshold logic

* switch to mutex instead of channel for better perf

* add back fix to remove double buffering

* rename local var for clarity

* revert cosmetic change

* Replace incoming frame channels with queues (#253)

* Replace incoming frame channels with queues

The rx channel in Session and link has been replaced with a queue. This
ensures that parents enqueueing frames to their children is always a non
blocking operation and removes the need to "pump the mux".
Consolidated some link creation code into newLink().
Links echoing a flow frame now correctly include session info.
Reverted workaround for sending disposition frames from senders as we
can now send directly from muxHandleFrame.

* minor cleanup from review

* Remove configuration of session window (#257)

The values weren't actually honored which is ok per spec, so remove them
from the public surface area.
Cleaned up handling of sender link credit.  Since we don't use
availableCredit, it's been removed.

* Improve handling of sender-settled messages (#258)

Reclaim credit for sender-settled messages when Prefetch or Receive are
called.  This removes the need to explicitly settle the message.