Make writes deterministic (#282)
* Make writes deterministic

Added Conn.WriteDeadline to control the write deadline when writing to
net.Conn.  The default value is 30 seconds.  This can be overridden by
APIs that take a context and the context has a deadline/timeout.
For APIs that directly send frames, wait for the frame to be written to
the underlying net.Conn before considering the write a success.

* add clarifying comments on Sent channel usage

* don't send if the deadline has exceeded

* don't allow cancellation of sending a frame

* remove TODO

* add TODO for receiver disposition ack timeout

* return transfer limit exceeded error for sender

* add another clarifying comment